refactor: remove bridge abstraction to MobileNumberValidator
